Extract vector index optimizer helpers into a separate file
