+++ ydb/core/kqp/opt/physical/effects/ya.make
@@ -6,12 +6,13 @@ SRCS(
     kqp_opt_phy_indexes.cpp
     kqp_opt_phy_insert_index.cpp
     kqp_opt_phy_insert.cpp
+    kqp_opt_phy_returning.cpp
     kqp_opt_phy_uniq_helper.cpp
     kqp_opt_phy_update_index.cpp
     kqp_opt_phy_update.cpp
-    kqp_opt_phy_upsert_index.cpp
-    kqp_opt_phy_returning.cpp
     kqp_opt_phy_upsert_defaults.cpp
+    kqp_opt_phy_upsert_index.cpp
+    kqp_opt_phy_vector_index.cpp
 )
 
 PEERDIR(
